% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip-methods.R
\name{addTips}
\alias{addTips}
\title{Add tips to a tree}
\usage{
addTips(tree, tids, sids, strt_ages = NULL, end_ages = 0,
  pids = paste0("p_", tids), progress = "none")
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{tids}{tip IDs}

\item{sids}{IDs of node that will become new tip sisters}

\item{strt_ages}{timepoints at which new tips first appear in the tree}

\item{end_ages}{timepoints at which new tips end appear in the tree, default 0.}

\item{pids}{parent ID (default is 'p_' + tid)}

\item{progress}{name of the progress bar to use, see \code{\link{create_progress_bar}}}
}
\description{
Returns a tree with tip ID(s) added
}
\details{
User must provide new tip ID(s), the ID(s) of the nodes
which will become the new tip's sister, and new branch lengths.
Optionally, user can specify the IDs for the new parent nodes.
Note, returned tree is no longer up-to-date.
}
\examples{
library(treeman)
tree <- randTree(10)
possible_ages <- getSpnAge(tree, 't1', tree['age'])
start_age <- runif(1, possible_ages[['end']], possible_ages[['start']])
end_age <- possible_ages[['end']]
tree <- addTips(tree, tids='t11', sids='t1', strt_ages=start_age,
end_ages=end_age)
tree <- updateTree(tree)
summary(tree)
}
\seealso{
\code{\link{rmTips}}, \code{\link{updateTree}},
\url{https://github.com/DomBennett/treeman/wiki/manip-methods}
}

