% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemen-declaration.R
\name{as-TreeMen}
\alias{as-TreeMen}
\title{Convert list to a TreeMen}
\description{
Return a \code{TreeMen} object from a list of \code{TreeMans}
}
\examples{
library(treeman)
trees <- list('tree_1'=randTree(10), 'tree_2'=randTree(10))
trees <- as(trees, 'TreeMen')
}
\seealso{
\code{\link{TreeMen-class}}
}

