% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-methods.R
\name{calcNodeBlnc}
\alias{calcNodeBlnc}
\title{Calculate the balance of a node}
\usage{
calcNodeBlnc(tree, id)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{id}{node id}
}
\description{
Returns the balance of a node.
}
\details{
Balance is calculated as the absolute difference between the number of descendents 
of the two bifurcating edges of a node and the expected value for a balanced tree.
\code{NA} is returned if the node is polytomous.
}
\examples{
library(treeman)
tree <- randTree(10)
calcNodeBlnc(tree, id=tree['root'])  # root balance
}
\seealso{
\code{\link{calcNodesBlnc}}, 
\url{https://github.com/DomBennett/treeman/wiki/calc-methods}
}

