% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-methods.R
\name{calcNodesBlnc}
\alias{calcNodesBlnc}
\title{Calculate the balances of all nodes}
\usage{
calcNodesBlnc(tree, ids, ...)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{...}{\code{plyr} arguments}

\item{id}{node id}
}
\description{
Returns the absolute differences in number of descendants for bifurcating 
branches of every node
}
\details{
Runs \code{calcNodeBlnc()} across all node IDs. \code{NA} is returned if the
node is polytomous. Parallelizable.
}
\examples{
library(treeman)
tree <- randTree(10)
calcNodesBlnc(tree, id=tree['nodes'])
}
\seealso{
\code{\link{calcNodeBlnc}}, 
\url{https://github.com/DomBennett/treeman/wiki/calc-methods}
}

