% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_methods.R
\name{checkTreeMan}
\alias{checkTreeMan}
\title{Check if tree is correct}
\usage{
checkTreeMan(object)
}
\arguments{
\item{object}{\code{TreeMan} object}
}
\description{
Return T/F if tree is a true \code{TreeMan} object
}
\details{
Tests whether each node in tree points to valid other node IDs. Also
ensures `spn` and `root` are correct. Reports nodes that have errors.
}
\seealso{
\code{\link{checkTreeMen}}
}

