% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_methods.R
\name{checkTreeMen}
\alias{checkTreeMen}
\title{Check if trees are correct}
\usage{
checkTreeMen(object)
}
\arguments{
\item{object}{\code{TreeMen} object}
}
\description{
Return T/F if trees is a true \code{TreeMen} object
}
\details{
Tests whether all trees in object are \code{TreeMan} objects
}
\seealso{
\code{\link{checkNdlst}}
}
