% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-methods.R
\name{downdateTree}
\alias{downdateTree}
\title{Downdate tree slots}
\usage{
downdateTree(tree)
}
\arguments{
\item{tree}{\code{TreeMan} object}
}
\description{
Return tree with memory heavy node slots removed.
}
\details{
Potential uses: reduce memory load by downdating a tree,
force get-methods to use ndlst rather than ndmtrx.
}
\examples{
# library(treeman)
tree <- randTree(10)
tree <- downdateTree(tree)
summary(tree)  # error
}
\seealso{
\code{\link{updateTree}}
}

