% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_methods.R
\name{fastCheckTreeMan}
\alias{fastCheckTreeMan}
\title{Check if tree is correct, fast!}
\usage{
fastCheckTreeMan(object)
}
\arguments{
\item{object}{\code{TreeMan} object}
}
\description{
Return T/F if tree is a true \code{TreeMan} object
}
\details{
Whenever a tree is first initiated this check is used.
For more detailed checking use \code{checkTreeMan}.
}
\seealso{
\code{\link{checkTreeMan}}, \code{\link{checkTreeMen}}
}

