% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getNdPrid}
\alias{getNdPrid}
\title{Get pre-nodes to root}
\usage{
getNdPrid(tree, id)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{id}{node id}
}
\description{
Return node ids for connecting \code{id} to root.
}
\details{
Returns a vector.
}
\examples{
library(treeman)
tree <- randTree(10)
# get all nodes to root
getNdPrid(tree, id='t1')
}
\seealso{
\code{\link{getNdsPrid}}, 
\code{\link{getNdPtid}}, 
\code{\link{getNdsPtid}}, 
\url{https://github.com/DomBennett/treeman/wiki/get-methods}
}

