% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getNdPtid}
\alias{getNdPtid}
\title{Get post-nodes to tips}
\usage{
getNdPtid(tree, id)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{id}{node id}
}
\description{
Return node ids for connecting \code{id} to kids.
}
\details{
Returns a vector.
}
\examples{
library(treeman)
tree <- randTree(10)
# get all nodes from root to tip
getNdPtid(tree, id='n1')
}
\seealso{
\code{\link{getNdsPtid}}, 
\code{\link{getNdPrid}}, 
\code{\link{getNdsPrid}}, 
\url{https://github.com/DomBennett/treeman/wiki/get-methods}
}

