% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getNdsMat}
\alias{getNdsMat}
\title{Get T/F matrix of connected node IDs}
\usage{
getNdsMat(tree, qry_ids)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{qry_ids}{vector of IDs}
}
\description{
Returns a TRUE/FALSE matrix of connected node IDs given
a vector of query IDs.
}
\details{
The function takes each ID in \code{qry_ids} and tests whether
each ID in tree descends from it. The resulting matrix has the \code{qry_ids}
as columns and all node IDs in tree as rows. Each column is a TRUE/FALSE vector
of prids, each row a TRUE/FALSE vector for ptids (when \code{qry_ids} equals all
node IDs). Kids can be calculated by node IDs using this function by providing
all tip IDs as \code{qry_ids}.

Functions that require running multiple get methods across all nodes may
run faster by using this function once. All \code{getNds_} use this function.
}
\examples{
library(treeman)
# generate tree and extract IDs
tree <- randTree(100)
all <- tree['all']
nids <- tree['nds']
tids <- tree['tips']
# get kids by only providing tip IDs
res <- getNdsMat(tree, tids)
kids <- apply(res, 1, function(x) tids[x])
# or count kids by node....
nkids <- apply(res, 1, sum)
# get ptids for internal nodes by slicing res
res <- getNdsMat(tree, all)
ptids <- apply(res, 1, function(x) all[x])
ptids <- ptids[nids]
# get prids by extracting columns
prids <- apply(res, 2, function(x) all[x])
}
\seealso{
\url{https://github.com/DomBennett/treeman/wiki/get-methods}
}

