% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getNdsPrdst}
\alias{getNdsPrdst}
\title{Get pre-distances}
\usage{
getNdsPrdst(tree, ids, ...)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{ids}{vector of node ids}

\item{...}{\code{plyr} arguments}
}
\description{
Return root to tip distances (prdst) for \code{ids}
}
\details{
Sums the lengths of all branches from \code{ids} to root.
}
\examples{
library(treeman)
tree <- randTree(10)
getNdsPrdst(tree, ids=tree['tips'])  # return prdsts for all tips
}
\seealso{
\code{\link{getNdPrdst}},
\url{https://github.com/DomBennett/treeman/wiki/get-methods}
}

