% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getNdsPrids}
\alias{getNdsPrids}
\title{Get pre-nodes for multiple nodes}
\usage{
getNdsPrids(tree, ids, ...)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{ids}{vector of node ids}

\item{...}{\code{plyr} arguments}
}
\description{
Return node ids for connecting \code{id} to root.
}
\details{
Returns a list, parallizable.
}
\examples{
library(treeman)
tree <- randTree(10)
getNdsPrids(tree, ids=tree['tips'])
}
\seealso{
\code{\link{getNdPrids}},
\code{\link{getNdPrid}},
\code{\link{getNdsPrid}}, 
\code{\link{getNdPtid}}, 
\code{\link{getNdsPtid}}, 
\url{https://github.com/DomBennett/treeman/wiki/get-methods}
}

