% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getNdsPtid}
\alias{getNdsPtid}
\title{Get post-nodes to tips for multiple nodes}
\usage{
getNdsPtid(tree, ids, ...)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{ids}{vector of node ids}

\item{...}{\code{plyr} arguments}
}
\description{
Return node ids for connecting \code{ids} to kids.
}
\details{
Returns a list, parallizable.
}
\examples{
library(treeman)
tree <- randTree(10)
# get all nodes to tip for all nodes
getNdsPtid(tree, ids=tree['nds'])
}
\seealso{
\code{\link{getNdPtid}}, 
\code{\link{getNdPrid}}, 
\code{\link{getNdsPrid}}, 
\url{https://github.com/DomBennett/treeman/wiki/get-methods}
}

