% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getNodeKids}
\alias{getNodeKids}
\title{Get children IDs}
\usage{
getNodeKids(tree, id)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{id}{node id}
}
\description{
Return the node ids of all tips that descend from node.
}
\details{
Returns a vector
}
\examples{
library(treeman)
tree <- randTree(10)
# everyone descends from root
getNodeKids(tree, id=tree['root'])
}
\seealso{
\code{\link{getNodesKid}}, 
\url{https://github.com/DomBennett/treeman/wiki/get-methods}
}

