% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getNodePrid}
\alias{getNodePrid}
\title{Get pre-nodes to root}
\usage{
getNodePrid(tree, id)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{id}{node id}
}
\description{
Return node ids for connecting \code{id} to root.
}
\details{
Returns a vector.
}
\examples{
library(treeman)
tree <- randTree(10)
# get all nodes to root
getNodePrid(tree, id='t1')
}
\seealso{
\code{\link{getNodesPrid}}, 
\code{\link{getNodePtid}}, 
\code{\link{getNodesPtid}}, 
\url{https://github.com/DomBennett/treeman/wiki/get-methods}
}

