% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getNodesAge}
\alias{getNodesAge}
\title{Get ages for multiple nodes}
\usage{
getNodesAge(tree, ids, ...)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{ids}{vector of node ids}

\item{...}{\code{plyr} arguments}
}
\description{
Return the root to tip distances for \code{ids}.
}
\details{
Returns a vector, parallelizable.
}
\examples{
library(treeman)
tree <- randTree(10)
getNodesAge(tree, ids=tree['nodes'])
}
\seealso{
\code{\link{getNodeAge}}, 
\code{\link{getSpanAge}}, 
\code{\link{getSpansAge}}, 
\url{https://github.com/DomBennett/treeman/wiki/get-methods}
}

