% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getNodesPtid}
\alias{getNodesPtid}
\title{Get post-nodes to tips for multiple nodes}
\usage{
getNodesPtid(tree, ids, ...)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{ids}{vector of node ids}

\item{...}{\code{plyr} arguments}
}
\description{
Return node ids for connecting \code{ids} to kids.
}
\details{
Returns a list, parallizable.
}
\examples{
library(treeman)
tree <- randTree(10)
# get all nodes to tip for all nodes
getNodesPtid(tree, ids=tree['nodes'])
}
\seealso{
\code{\link{getNodePtid}}, 
\code{\link{getNodePrid}}, 
\code{\link{getNodesPrid}}, 
\url{https://github.com/DomBennett/treeman/wiki/get-methods}
}

