% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getNodesSister}
\alias{getNodesSister}
\title{Get sister id}
\usage{
getNodesSister(tree, ids, ...)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{name}{slot name}

\item{id}{node id}
}
\description{
Returns the ids of the sister(s) of node ids given.
}
\details{
An error is raised if there is no sister (e.g. for the root).
 There can be more than one sister if tree is polytomous.
}
\examples{
library(treeman)
tree <- randTree(10)
getNodesSister(tree, ids=tree['tips'])
}
\seealso{
\code{\link{getNodeSister}}, 
\url{https://github.com/DomBennett/treeman/wiki/get-methods}
}

