% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getNodesSlot}
\alias{getNodesSlot}
\title{Get a node slot for multiple nodes}
\usage{
getNodesSlot(tree, name, ids, ...)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{name}{slot name}

\item{ids}{vector of node ids}

\item{...}{\code{plyr} arguments}
}
\description{
Returns the value of named slot.
}
\details{
Returned object depends on name, either character, vector or numeric. Parallelizable.
}
\examples{
library(treeman)
tree <- randTree(10)
getNodesSlot(tree, name='span', ids=tree['tips'])  # return spans of all tips
}
\seealso{
\code{\link{getNodeSlot}}, 
\url{https://github.com/DomBennett/treeman/wiki/get-methods}
}

