% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getSpanAge}
\alias{getSpanAge}
\title{Get age range}
\usage{
getSpanAge(tree, id)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{id}{node id}
}
\description{
Return start and end root to tip distances for \code{id}.
}
\details{
Returns a dataframe.
}
\examples{
library(treeman)
data(mammals)
getSpanAge(mammals, id='Homo_sapiens')
}
\seealso{
\code{\link{getNodeAge}}, 
\code{\link{getNodesAge}}, 
\code{\link{getSpansAge}}, 
\url{https://github.com/DomBennett/treeman/wiki/get-methods}
}

