% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getSpansAge}
\alias{getSpansAge}
\title{Get age ranges for multiple nodes}
\usage{
getSpansAge(tree, ids, ...)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{ids}{vector of node ids}

\item{...}{\code{plyr} arguments}
}
\description{
Return start and end root to tip distances for \code{ids}.
}
\details{
Returns a dataframe, parallelizable.
}
\examples{
library(treeman)
tree <- randTree(10)
# all nodes but root
ids <- tree['nodes'][tree['nodes'] != tree['root']]
getSpansAge(tree, ids=ids)
}
\seealso{
\code{\link{getNodeAge}}, 
\code{\link{getNodesAge}}, 
\code{\link{getSpanAge}}, 
\url{https://github.com/DomBennett/treeman/wiki/get-methods}
}

