% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getSpnAge}
\alias{getSpnAge}
\title{Get age range}
\usage{
getSpnAge(tree, id)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{id}{node id}
}
\description{
Return start and end root to tip distances for \code{id}.
}
\details{
Returns a dataframe.
}
\examples{
library(treeman)
data(mammals)
getSpnAge(mammals, id='Homo_sapiens')
}
\seealso{
\code{\link{getNdAge}}, 
\code{\link{getNdsAge}}, 
\code{\link{getSpnsAge}}, 
\url{https://github.com/DomBennett/treeman/wiki/get-methods}
}

