% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-spcl-methods.R
\name{getSubtree}
\alias{getSubtree}
\title{Get subtree}
\usage{
getSubtree(tree, id)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{id}{node id}
}
\description{
Return tree descending from \code{id}.
}
\details{
Returns a \code{TreeMan}, parallelizable. \code{id} must be an internal node.
}
\examples{
library(treeman)
data(mammals)
# get tree of apes
ape_id <- getPrnt(mammals, ids=c('Homo_sapiens', 'Hylobates_concolor'))
apes <- getSubtree(mammals, id=ape_id)
summary(apes)
}
\seealso{
\code{\link{getPrnt}}, \code{\link{addClade}},
\url{https://github.com/DomBennett/treeman/wiki/get-methods}
}
