% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getTxnyms}
\alias{getTxnyms}
\title{Get node id(s) for txonyms}
\usage{
getTxnyms(tree, txnyms, ...)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{txnyms}{vector of taxonomic names}

\item{...}{\code{plyr} arguments}
}
\description{
Returns the node ids of nodes with given taxonyms.
}
\details{
Returns a \code{list}, parallelizable.
}
\examples{
library(treeman)
data(mammals)
homo_ids <- getTxnyms(mammals, txnyms='Homo')
}
\seealso{
\code{\link{getNdLng}}, 
\code{\link{getNdsLng}}, 
\url{https://github.com/DomBennett/treeman/wiki/get-methods}
}

