% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node-declaration.R
\docType{class}
\name{Node}
\alias{Node}
\title{S4 class for displaying nodes}
\description{
The \code{Node} class is used to display node information.
It is only generated when a user implements the \code{[[]]} on a tree.
}
\section{Slots}{

\describe{
\item{\code{id}}{unique ID for node in tree['ndlst']}

\item{\code{spn}}{length of preceding branch}

\item{\code{prid}}{parent node ID}

\item{\code{ptid}}{child node ID}

\item{\code{kids}}{descending tip IDs}

\item{\code{nkids}}{number of descending tip IDs}

\item{\code{txnym}}{list of associated taxonyms}

\item{\code{pd}}{total branch length represented by node}

\item{\code{prdst}}{total branch length of connected prids}

\item{\code{age}}{age of node in tree}

\item{\code{root}}{T/F root node?}

\item{\code{tip}}{T/F tip node?}
}}
\seealso{
\code{\link{cTrees}}
}

