% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeman-declaration.R
\docType{methods}
\name{viz}
\alias{viz}
\alias{viz,TreeMan-method}
\title{Method viz}
\usage{
viz(tree, taxonyms = FALSE)

\S4method{viz}{TreeMan}(tree, taxonyms = FALSE)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{taxonyms}{Boolean, show taxonyms rather than IDs?}
}
\description{
Crude plot of \code{TreeMan} objects
}
\seealso{
\code{\link{TreeMan-class}}
}

