\name{treeman-package}
\alias{treeman-package}
\alias{treeman}
\docType{package}
\title{
Phylogenetic Tree Manipulation
}
\description{
Manipulate phylogenetic trees in R simply, intuitively and efficiently with
a list-based tree structure.
}
\details{
\tabular{ll}{
Package: \tab treeman\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-12-02\cr
License: \tab GPL-2\cr
}
I will do this later.
}
\author{
D.J. Bennett

Maintainer: D.J. Bennett <dominic.john.bennett@gmail.com>
}
\references{
No key references
}
phylogeny
evolution
macroevolution
tree simulation
\keyword{ package }
\seealso{
}
\examples{
}
