% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvrt-methods.R
\name{TreeMan-to-phylo}
\alias{TreeMan-to-phylo}
\title{Convert TreeMan to phylo}
\description{
Return ape's \code{phylo} from a \code{TreeMan}
}
\examples{
library(treeman)
library(ape)
tree <- randTree(10)
tree <- as(tree, 'phylo')
}
\seealso{
\code{\link{phylo-to-TreeMan}},
\code{\link{TreeMan-class}}
}

