% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemen-declaration.R
\docType{class}
\name{TreeMen-class}
\alias{TreeMen-class}
\alias{TreeMen-method}
\alias{[,TreeMen,character-method}
\alias{[[,TreeMen,ANY-method}
\alias{as.character,TreeMen-method}
\alias{cTrees,TreeMen-method}
\alias{print,TreeMen-method}
\alias{show,TreeMen-method}
\alias{str,TreeMen-method}
\title{TreeMen-class}
\usage{
\S4method{cTrees}{TreeMen}(x, ...)

\S4method{[[}{TreeMen,ANY}(x, i)

\S4method{[}{TreeMen,character}(x, i)

\S4method{as.character}{TreeMen}(x)

\S4method{show}{TreeMen}(object)

\S4method{str}{TreeMen}(object, max.level = 2L, ...)

\S4method{print}{TreeMen}(x)
}
\arguments{
\item{x}{\code{TreeMen} object}

\item{...}{additional tree objects}

\item{i}{tree index (integer or character)}

\item{object}{\code{TreeMen} object}

\item{max.level}{\code{str()} maximum level}
}
\description{
S4 class for multiple phylogenetic trees
}
\section{Slots}{

\describe{
\item{\code{treelist}}{list of \code{TreeMan} objects}

\item{\code{ntips}}{sum of tips per tree}

\item{\code{ntrees}}{total number of trees}
}}
\seealso{
\code{\link{cTrees}}
}

