% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemen-declaration.R
\docType{class}
\name{TreeMen}
\alias{TreeMen}
\title{S4 class for multiple phylogenetic trees}
\section{Slots}{

\describe{
\item{\code{treelist}}{list of \code{TreeMan} objects}

\item{\code{ntips}}{sum of tips per tree}

\item{\code{ntrees}}{total number of trees}
}}
\seealso{
\code{\link{cTrees}}
}

