% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeman-declaration.R
\name{viz}
\alias{viz}
\title{Visualise a tree}
\usage{
viz(tree, taxonyms = FALSE)
}
\arguments{
\item{tree}{\code{TreeMan}}

\item{taxonyms}{T/F, show the nodes with their taxonyms instead of IDs}
}
\description{
Crude plot of \code{TreeMan} objects
}

