% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip-methods.R
\name{addTips}
\alias{addTips}
\title{Add tips to a tree}
\usage{
addTips(tree, tids, sids, strt_ages = NULL, end_ages = 0,
  pids = paste0("p_", tids), progress = "none")
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{tids}{tip IDs}

\item{sids}{IDs of node that will become new tip sisters}

\item{pids}{parent ID (default is 'p_' + tid)}

\item{progress}{name of the progress bar to use, see \code{\link{create_progress_bar}}}

\item{spns}{length of new branch}
}
\description{
Returns a tree with tip ID(s) added
}
\details{
User must provide new tip ID(s), the ID(s) of the nodes
which will become the new tip's sister, and new branch lengths.
Optionally, user can specify the IDs for the new parent nodes.
Note, returned tree is no longer up-to-date.
}
\examples{
library(treeman)
tree <- randTree(10)
tree <- addTips(tree, tids='t11', sids='t1', spns=1)
tree <- updateTree(tree)
summary(tree)
}
\seealso{
\code{\link{rmTips}}, \code{\link{updateTree}},
\url{https://github.com/DomBennett/treeman/wiki/manip-methods}
}

