% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getNdsPrids}
\alias{getNdsPrids}
\title{Get pre-nodes for multiple nodes}
\usage{
getNdsPrids(tree, ids, parallel = FALSE, progress = "none")
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{ids}{vector of node ids}

\item{parallel}{logical, make parallel?}

\item{progress}{name of the progress bar to use, see \code{\link{create_progress_bar}}}
}
\description{
Return node ids for connecting \code{id} to root.
}
\details{
Returns a list, parallizable. Order of prids may vary.
}
\examples{
library(treeman)
tree <- randTree(10)
getNdsPrids(tree, ids=tree['tips'])
}
\seealso{
\code{\link{getNdPrids}},
\code{\link{getNdPtids}}, 
\code{\link{getNdsPtids}}, 
\url{https://github.com/DomBennett/treeman/wiki/get-methods}
}

