% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-spcl-methods.R
\name{getTreeAge}
\alias{getTreeAge}
\title{Get age of tree}
\usage{
getTreeAge(tree, parallel = FALSE)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{parallel}{logical, make parallel?}
}
\description{
Returns age, numeric, of tree
}
\details{
This can also be achieved with \code{tree['age']} but will
only work if the the tree has been updated with \code{updateTree()}.
For faster computation, especially within function that perform multiple
tree manipulations where the whole tree doesn't need updating, use this function.
Parallelizable.
}
\examples{
library(treeman)
tree <- randTree(10)
(getTreeAge(tree))
}
\seealso{
\code{\link{updateTree}}, 
\url{https://github.com/DomBennett/treeman/wiki/get-methods}
}

