% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip-methods.R
\name{rmTips}
\alias{rmTips}
\title{Remove tips from a tree}
\usage{
rmTips(tree, tid, drp_intrnl = TRUE)
}
\arguments{
\item{tree}{\code{TreeMan} object}

\item{tid}{tip ID}

\item{drp_intrnl}{Boolean, drop internal branches, default FALSE
\url{https://github.com/DomBennett/treeman/wiki/manip-methods}}
}
\description{
Returns a tree with a tip ID(s) for removal
}
\details{
Removes tips in a tree. Set drp_intrnl to FALSE to convert
internal nodes into new tips.
}
\examples{
library(treeman)
tree <- randTree(10)
tree <- rmTips(tree, 't1')
tree <- updateTree(tree)
summary(tree)
}

