% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-methods.R
\name{updateTree}
\alias{updateTree}
\title{Update tree slots after manipulation}
\usage{
updateTree(tree)
}
\arguments{
\item{tree}{\code{TreeMan} object}
}
\description{
Return tree with updated node slots.
}
\details{
Tree slots in the \code{TreeMan} object are not automatically updated
to make mainipulations computationally faster. To update these slots, run this function.
}
\seealso{
\code{\link{downdateTree}}, \code{\link{getTreeAge}}
}

